+++ text_speech/bark/generation_utils.py
@@ -3,10 +3,9 @@ import re
 
 import logging
 import numpy as np
-from scipy.special import softmax
-import torch
-import torch.nn.functional as F
 import tqdm
+
+from math_utils import softmax
 
 CONTEXT_WINDOW_SIZE = 1024
 
@@ -44,94 +43,26 @@ for _, lang in SUPPORTED_LANGS:
 
 logger = logging.getLogger(__name__)
 
-CUR_PATH = os.path.dirname(os.path.abspath(__file__))
+this_path = os.path.dirname(os.path.abspath(__file__))
 
-default_cache_dir = os.path.join(os.path.expanduser("~"), ".cache")
-CACHE_DIR = os.path.join(os.getenv("XDG_CACHE_HOME", default_cache_dir), "suno", "bark_v0")
-
-
-def _cast_bool_env_var(s):
-    return s.lower() in ('true', '1', 't')
-
-
-USE_SMALL_MODELS = _cast_bool_env_var(os.environ.get("SUNO_USE_SMALL_MODELS", "False"))
-GLOBAL_ENABLE_MPS = _cast_bool_env_var(os.environ.get("SUNO_ENABLE_MPS", "False"))
-OFFLOAD_CPU = _cast_bool_env_var(os.environ.get("SUNO_OFFLOAD_CPU", "False"))
-
-REMOTE_MODEL_PATHS = {
-    "text_small": {
-        "repo_id": "suno/bark",
-        "file_name": "text.pt",
-    },
-    "coarse_small": {
-        "repo_id": "suno/bark",
-        "file_name": "coarse.pt",
-    },
-    "fine_small": {
-        "repo_id": "suno/bark",
-        "file_name": "fine.pt",
-    },
-    "text": {
-        "repo_id": "suno/bark",
-        "file_name": "text_2.pt",
-    },
-    "coarse": {
-        "repo_id": "suno/bark",
-        "file_name": "coarse_2.pt",
-    },
-    "fine": {
-        "repo_id": "suno/bark",
-        "file_name": "fine_2.pt",
-    },
-}
-
-if not hasattr(torch.nn.functional, 'scaled_dot_product_attention') and torch.cuda.is_available():
-    logger.warning(
-        "torch version does not support flash attention. You will get faster" +
-        " inference speed by upgrade torch to newest nightly version."
-    )
-
-
-class InferenceContext:
-    def __init__(self, benchmark=False):
-        # we can't expect inputs to be the same length, so disable benchmarking by default
-        self._chosen_cudnn_benchmark = benchmark
-        self._cudnn_benchmark = None
-
-    def __enter__(self):
-        self._cudnn_benchmark = torch.backends.cudnn.benchmark
-        torch.backends.cudnn.benchmark = self._chosen_cudnn_benchmark
-
-    def __exit__(self, exc_type, exc_value, exc_traceback):
-        torch.backends.cudnn.benchmark = self._cudnn_benchmark
-
-
-if torch.cuda.is_available():
-    torch.backends.cuda.matmul.allow_tf32 = True
-    torch.backends.cudnn.allow_tf32 = True
-
+onnx = False
 
 ####
 # Generation Functionality
 ####
 
 
-def _tokenize(tokenizer, text):
-    return tokenizer.encode(text, add_special_tokens=False)
-
-
-def _detokenize(tokenizer, enc_text):
-    return tokenizer.decode(enc_text)
-
-
-def _normalize_whitespace(text):
-    return re.sub(r"\s+", " ", text).strip()
-
-
 TEXT_ENCODING_OFFSET = 10_048
 SEMANTIC_PAD_TOKEN = 10_000
 TEXT_PAD_TOKEN = 129_595
 SEMANTIC_INFER_TOKEN = 129_599
+
+COARSE_SEMANTIC_PAD_TOKEN = 12_048
+COARSE_INFER_TOKEN = 12_050
+
+
+def _normalize_whitespace(text):
+    return re.sub(r"\s+", " ", text).strip()
 
 
 def _load_history_prompt(history_prompt_input):
@@ -143,7 +74,7 @@ def _load_history_prompt(history_prompt_input):
         if history_prompt_input not in ALLOWED_PROMPTS:
             raise ValueError("history prompt not found")
         history_prompt = np.load(
-            os.path.join(CUR_PATH, "assets", "prompts", f"{history_prompt_input}.npz")
+            os.path.join(this_path, "assets", "prompts", f"{history_prompt_input}.npz")
         )
     elif isinstance(history_prompt_input, dict):
         assert ("semantic_prompt" in history_prompt_input)
@@ -164,20 +95,116 @@ def _flatten_codebooks(arr, offset_size=CODEBOOK_SIZE):
     return flat_arr
 
 
-COARSE_SEMANTIC_PAD_TOKEN = 12_048
-COARSE_INFER_TOKEN = 12_050
+def generate_text_semantic(
+        models,
+        text,
+        temp=0.7,
+        silent=False,
+        min_eos_p=0.2,
+        max_gen_duration_s=None,
+        allow_early_stop=True):
+    """Generate semantic tokens from text."""
+    text = _normalize_whitespace(text)
+    assert len(text.strip()) > 0
+
+    tokenizer = models["tokenizer"]
+    encoded_text = np.array(tokenizer.encode(text, add_special_tokens=False))
+    encoded_text = encoded_text + TEXT_ENCODING_OFFSET
+
+    if len(encoded_text) > 256:
+        p = round((len(encoded_text) - 256) / len(encoded_text) * 100, 1)
+        logger.warning(f"warning, text too long, lopping of last {p}%")
+        encoded_text = encoded_text[:256]
+
+    encoded_text = np.pad(
+        encoded_text,
+        (0, 256 - len(encoded_text)),
+        constant_values=TEXT_PAD_TOKEN,
+        mode="constant",
+    )
+
+    semantic_history = np.array([SEMANTIC_PAD_TOKEN] * 256)
+    x = np.hstack([
+        encoded_text, semantic_history, np.array([SEMANTIC_INFER_TOKEN])
+    ]).astype(np.int64)
+    x = np.expand_dims(x, axis=0)
+    assert x.shape[1] == 256 + 256 + 1
+
+    net = models["net"]
+    kv_cache = None
+
+    pbar_state = 0
+    tot_generated_duration_s = 0
+    n_tot_steps = 768
+    pbar = tqdm.tqdm(disable=silent, total=n_tot_steps)
+    for n in range(n_tot_steps):
+        if kv_cache is not None:
+            x_input = x[:, [-1]]
+        else:
+            x_input = x
+            kv_cache = np.zeros((48, 16, 0, 64), dtype=np.float32)
+
+        # feedforward
+        if not onnx:
+            output = net.predict([
+                x_input, kv_cache,
+            ])
+        else:
+            output = net.run(None, {
+                'x_input': x_input, 'past_kv': kv_cache,
+            })
+        logits, kv_cache = output
+
+        relevant_logits = logits[0, 0, :SEMANTIC_VOCAB_SIZE]
+        if allow_early_stop:
+            relevant_logits = np.hstack(
+                (relevant_logits, logits[0, 0, [SEMANTIC_PAD_TOKEN]])  # eos
+            )
+
+        probs = softmax(relevant_logits / temp, axis=-1).astype(np.float64)
+        item_next = np.random.multinomial(1, probs / sum(probs))
+        item_next = np.argsort(-item_next)[:1]
+
+        if allow_early_stop and (
+                item_next == SEMANTIC_VOCAB_SIZE
+                or (min_eos_p is not None and probs[-1] >= min_eos_p)
+        ):
+            # eos found, so break
+            pbar.update(n - pbar_state)
+            break
+
+        x = np.concatenate((x, item_next[None]), axis=1)
+        tot_generated_duration_s += 1 / SEMANTIC_RATE_HZ
+        if max_gen_duration_s is not None \
+                and tot_generated_duration_s > max_gen_duration_s:
+            pbar.update(n - pbar_state)
+            break
+        if n == n_tot_steps - 1:
+            pbar.update(n - pbar_state)
+            break
+
+        if n > pbar_state:
+            if n > pbar.total:
+                pbar.total = n
+            pbar.update(n - pbar_state)
+
+        pbar_state = n
+
+    pbar.total = n
+    pbar.refresh()
+    pbar.close()
+    out = x.squeeze()[256 + 256 + 1:]
+
+    return out
 
 
 def generate_coarse(
+        models,
         x_semantic,
         history_prompt=None,
         temp=0.7,
-        top_k=None,
-        top_p=None,
         max_coarse_history=630,  # min 60 (faster), max 630 (more context)
-        sliding_window_len=60,
-        use_kv_caching=False,
-):
+        sliding_window_len=60):
     """Generate coarse audio codes from semantic tokens."""
     semantic_to_coarse_ratio = COARSE_RATE_HZ / SEMANTIC_RATE_HZ * N_COARSE_CODEBOOKS
     max_semantic_history = int(np.floor(max_coarse_history / semantic_to_coarse_ratio))
@@ -204,6 +231,8 @@ def generate_coarse(
         x_semantic_history = np.array([], dtype=np.int32)
         x_coarse_history = np.array([], dtype=np.int32)
 
+    net = models["coarse"]
+
     # start loop
     n_steps = int(
         round(
@@ -224,64 +253,57 @@ def generate_coarse(
         # pad from right side
         x_in = x_semantic_in[:, np.max([0, semantic_idx - max_semantic_history]):]
         x_in = x_in[:, :256]
-        x_in = F.pad(
+        x_in = np.pad(
             x_in,
-            (0, 256 - x_in.shape[-1]),
+            [(0, 0), (0, 256 - x_in.shape[-1])],
             "constant",
-            COARSE_SEMANTIC_PAD_TOKEN,
+            constant_values=COARSE_SEMANTIC_PAD_TOKEN,
         )
-        x_in = np.hstack(
-            [
-                x_in,
-                torch.tensor([COARSE_INFER_TOKEN])[None],
-                x_coarse_in[:, -max_coarse_history:],
-            ]
-        )
+        x_in = np.hstack([
+            x_in,
+            np.array([COARSE_INFER_TOKEN])[None],
+            x_coarse_in[:, -max_coarse_history:],
+        ])
+
         kv_cache = None
         for _ in range(sliding_window_len):
             if n_step >= n_steps:
                 continue
             is_major_step = n_step % N_COARSE_CODEBOOKS == 0
 
-            if use_kv_caching and kv_cache is not None:
+            if kv_cache is not None:
                 x_input = x_in[:, [-1]]
             else:
                 x_input = x_in
+                kv_cache = np.zeros((48, 16, 0, 64), dtype=np.float32)
 
-            # print("model---", x_input.shape)
-            # print("kv_cache---", (len(kv_cache), len(kv_cache[0]), kv_cache[0][0].shape) if kv_cache is not None else None)
-            logits, kv_cache = model(x_input, use_cache=use_kv_caching, past_kv=kv_cache)
-            logit_start_idx = (
-                    SEMANTIC_VOCAB_SIZE + (1 - int(is_major_step)) * CODEBOOK_SIZE
-            )
-            logit_end_idx = (
-                    SEMANTIC_VOCAB_SIZE + (2 - int(is_major_step)) * CODEBOOK_SIZE
-            )
+            # feedforward
+            if not onnx:
+                output = net.predict([
+                    x_input, kv_cache,
+                ])
+            else:
+                output = net.run(None, {
+                    'x_input': x_input, 'past_kv': kv_cache,
+                })
+            logits, kv_cache = output
+
+            logit_start_idx = \
+                SEMANTIC_VOCAB_SIZE + (1 - int(is_major_step)) * CODEBOOK_SIZE
+            logit_end_idx = \
+                SEMANTIC_VOCAB_SIZE + (2 - int(is_major_step)) * CODEBOOK_SIZE
+
             relevant_logits = logits[0, 0, logit_start_idx:logit_end_idx]
-            if top_p is not None:
-                # faster to convert to numpy
-                original_device = relevant_logits.device
-                relevant_logits = relevant_logits.detach().cpu().type(torch.float32).numpy()
-                sorted_indices = np.argsort(relevant_logits)[::-1]
-                sorted_logits = relevant_logits[sorted_indices]
-                cumulative_probs = np.cumsum(softmax(sorted_logits))
-                sorted_indices_to_remove = cumulative_probs > top_p
-                sorted_indices_to_remove[1:] = sorted_indices_to_remove[:-1].copy()
-                sorted_indices_to_remove[0] = False
-                relevant_logits[sorted_indices[sorted_indices_to_remove]] = -np.inf
-                relevant_logits = torch.from_numpy(relevant_logits)
-                relevant_logits = relevant_logits.to(original_device)
-            if top_k is not None:
-                v, _ = torch.topk(relevant_logits, min(top_k, relevant_logits.size(-1)))
-                relevant_logits[relevant_logits < v[-1]] = -float("Inf")
-            probs = F.softmax(relevant_logits / temp, dim=-1)
-            item_next = torch.multinomial(probs, num_samples=1).to(torch.int32)
+
+            probs = softmax(relevant_logits / temp, axis=-1).astype(np.float64)
+            item_next = np.random.multinomial(1, probs / sum(probs))
+            item_next = np.argsort(-item_next)[:1]
             item_next += logit_start_idx
-            x_coarse_in = torch.cat((x_coarse_in, item_next[None]), dim=1)
-            x_in = torch.cat((x_in, item_next[None]), dim=1)
-            del logits, relevant_logits, probs, item_next
+
+            x_coarse_in = np.concatenate((x_coarse_in, item_next[None]), axis=1)
+            x_in = np.concatenate((x_in, item_next[None]), axis=1)
+
             n_step += 1
-        del x_in
 
     gen_coarse_arr = x_coarse_in[len(x_coarse_history):]
     gen_coarse_audio_arr = gen_coarse_arr.reshape(-1, N_COARSE_CODEBOOKS).T - SEMANTIC_VOCAB_SIZE
@@ -292,43 +314,21 @@ def generate_coarse(
 
 
 def generate_fine(
+        models,
         x_coarse_gen,
         history_prompt=None,
         temp=0.5,
-        silent=True,
-):
+        silent=True):
     """Generate full audio codes from coarse audio codes."""
-    assert (
-            isinstance(x_coarse_gen, np.ndarray)
-            and len(x_coarse_gen.shape) == 2
-            and 1 <= x_coarse_gen.shape[0] <= N_FINE_CODEBOOKS - 1
-            and x_coarse_gen.shape[1] > 0
-            and x_coarse_gen.min() >= 0
-            and x_coarse_gen.max() <= CODEBOOK_SIZE - 1
-    )
     if history_prompt is not None:
         history_prompt = _load_history_prompt(history_prompt)
         x_fine_history = history_prompt["fine_prompt"]
-        assert (
-                isinstance(x_fine_history, np.ndarray)
-                and len(x_fine_history.shape) == 2
-                and x_fine_history.shape[0] == N_FINE_CODEBOOKS
-                and x_fine_history.shape[1] >= 0
-                and x_fine_history.min() >= 0
-                and x_fine_history.max() <= CODEBOOK_SIZE - 1
-        )
     else:
         x_fine_history = None
     n_coarse = x_coarse_gen.shape[0]
-    # load models if not yet exist
-    global models
-    global models_devices
-    if "fine" not in models:
-        preload_models()
+
     model = models["fine"]
-    if OFFLOAD_CPU:
-        model.to(models_devices["fine"])
-    device = next(model.parameters()).device
+
     # make input arr
     in_arr = np.vstack(
         [
@@ -361,42 +361,39 @@ def generate_fine(
         )
     # we can be lazy about fractional loop and just keep overwriting codebooks
     n_loops = np.max([0, int(np.ceil((x_coarse_gen.shape[1] - (1024 - n_history)) / 512))]) + 1
-    with _inference_mode():
-        in_arr = torch.tensor(in_arr.T).to(device)
-        for n in tqdm.tqdm(range(n_loops), disable=silent):
-            start_idx = np.min([n * 512, in_arr.shape[0] - 1024])
-            start_fill_idx = np.min([n_history + n * 512, in_arr.shape[0] - 512])
-            rel_start_fill_idx = start_fill_idx - start_idx
-            in_buffer = in_arr[start_idx: start_idx + 1024, :][None]
-            for nn in range(n_coarse, N_FINE_CODEBOOKS):
-                logits = model(nn, in_buffer)
-                if temp is None:
-                    relevant_logits = logits[0, rel_start_fill_idx:, :CODEBOOK_SIZE]
-                    codebook_preds = torch.argmax(relevant_logits, -1)
-                else:
-                    relevant_logits = logits[0, :, :CODEBOOK_SIZE] / temp
-                    probs = F.softmax(relevant_logits, dim=-1)
-                    codebook_preds = torch.multinomial(
-                        probs[rel_start_fill_idx:1024], num_samples=1
-                    ).reshape(-1)
-                codebook_preds = codebook_preds.to(torch.int32)
-                in_buffer[0, rel_start_fill_idx:, nn] = codebook_preds
-                del logits, codebook_preds
-            # transfer over info into model_in and convert to numpy
-            for nn in range(n_coarse, N_FINE_CODEBOOKS):
-                in_arr[
-                start_fill_idx: start_fill_idx + (1024 - rel_start_fill_idx), nn
-                ] = in_buffer[0, rel_start_fill_idx:, nn]
-            del in_buffer
-        gen_fine_arr = in_arr.detach().cpu().numpy().squeeze().T
-        del in_arr
-    if OFFLOAD_CPU:
-        model.to("cpu")
+
+    in_arr = torch.tensor(in_arr.T)
+    for n in tqdm.tqdm(range(n_loops), disable=silent):
+        start_idx = np.min([n * 512, in_arr.shape[0] - 1024])
+        start_fill_idx = np.min([n_history + n * 512, in_arr.shape[0] - 512])
+        rel_start_fill_idx = start_fill_idx - start_idx
+        in_buffer = in_arr[start_idx: start_idx + 1024, :][None]
+        for nn in range(n_coarse, N_FINE_CODEBOOKS):
+            logits = model(nn, in_buffer)
+            if temp is None:
+                relevant_logits = logits[0, rel_start_fill_idx:, :CODEBOOK_SIZE]
+                codebook_preds = torch.argmax(relevant_logits, -1)
+            else:
+                relevant_logits = logits[0, :, :CODEBOOK_SIZE] / temp
+                probs = F.softmax(relevant_logits, dim=-1)
+                codebook_preds = torch.multinomial(
+                    probs[rel_start_fill_idx:1024], num_samples=1
+                ).reshape(-1)
+            codebook_preds = codebook_preds.to(torch.int32)
+            in_buffer[0, rel_start_fill_idx:, nn] = codebook_preds
+            del logits, codebook_preds
+        # transfer over info into model_in and convert to numpy
+        for nn in range(n_coarse, N_FINE_CODEBOOKS):
+            in_arr[
+            start_fill_idx: start_fill_idx + (1024 - rel_start_fill_idx), nn
+            ] = in_buffer[0, rel_start_fill_idx:, nn]
+        del in_buffer
+    gen_fine_arr = in_arr.detach().cpu().numpy().squeeze().T
+
     gen_fine_arr = gen_fine_arr[:, n_history:]
     if n_remove_from_end > 0:
         gen_fine_arr = gen_fine_arr[:, :-n_remove_from_end]
-    assert gen_fine_arr.shape[-1] == x_coarse_gen.shape[-1]
-    _clear_cuda_cache()
+
     return gen_fine_arr
